[XABT] Refactor manifest merging and ACW map generation out of `GenerateJavaStubs`.
