+++ src/Xamarin.Android.Build.Tasks/Tasks/GeneratePackageManagerJava.cs
@@ -44,6 +44,9 @@ namespace Xamarin.Android.Tasks
 		public string EnvironmentOutputDirectory { get; set; }
 
 		[Required]
+		public string IntermediateOutputDirectory { get; set; } = "";
+
+		[Required]
 		public string MainAssembly { get; set; }
 
 		[Required]
@@ -309,7 +312,7 @@ namespace Xamarin.Android.Tasks
 			ConcurrentDictionary<AndroidTargetArch, NativeCodeGenState>? nativeCodeGenStates = null;
 			if (enableMarshalMethods) {
 				nativeCodeGenStates = BuildEngine4.GetRegisteredTaskObjectAssemblyLocal<ConcurrentDictionary<AndroidTargetArch, NativeCodeGenState>> (
-					ProjectSpecificTaskObjectKey (GenerateJavaStubs.NativeCodeGenStateRegisterTaskKey),
+					MonoAndroidHelper.GetProjectBuildSpecificTaskObjectKey (GenerateJavaStubs.NativeCodeGenStateRegisterTaskKey, WorkingDirectory, IntermediateOutputDirectory),
 					RegisteredTaskObjectLifetime.Build
 				);
 			}
